sync the rest of the code with previous header changes
* exec/coroipcs.c (coroipcs_response_send)
(coroipcs_dispatch_send):
* exec/coroipcs.h (handler_fn_get):
* include/corosync/cpg.h (cpg_deliver_fn_t, cpg_confchg_fn_t):
* test/cpgbench.c (cpg_bm_confchg_fn, cpg_bm_deliver_fn):
* test/testcpg.c (print_cpgname, DeliverCallback)
(ConfchgCallback):
* test/testcpg2.c (deliver, confch):

git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@2044 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ exec/coroipcs.c
@@ -584,7 +584,7 @@ void *coroipcs_private_data_get (void *conn)
 	return (conn_info->private_data);
 }
 
-int coroipcs_response_send (void *conn, const void *msg, int mlen)
+int coroipcs_response_send (void *conn, const void *msg, size_t mlen)
 {
 	struct conn_info *conn_info = (struct conn_info *)conn;
 	struct sembuf sop;
@@ -865,7 +865,7 @@ void coroipcs_refcount_dec (void *conn)
 	pthread_mutex_unlock (&conn_info->mutex);
 }
 
-int coroipcs_dispatch_send (void *conn, const void *msg, int mlen)
+int coroipcs_dispatch_send (void *conn, const void *msg, size_t mlen)
 {
 	struct iovec iov;
